requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary>=2.9
